Renaming files for consistency
