+++ skrub/_drop_null_column.py
